requirements-freeze.txt
anyio==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
db-dtypes==1.1.1
debugpy==1.8.14
decorator==5.1.1
defusedxml==0.7.1
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.57.0
fqdn==1.5.1
google-api-core==2.11.1
google-auth==2.17.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.4
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-dataflow-client==0.8.4
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.0.13
jupyterlab_pygments==0.3.0
jupyterlab_server==2.24.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
MarkupSafe==2.1.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.0.8
notebook_shim==0.2.4
numpy==1.24.2
oauthlib==3.2.2
overrides==6.5.0
packaging==23.0
pandas==1.5.3
pandas-gbq==0.19.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydata-google-auth==1.8.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==3.3.0
pytz==2023.3
PyYAML==6.0
pyzmq==26.4.0
referencing==0.30.2
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
rsa==4.9
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.7
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.5.0
uri-template==1.3.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.14
zipp==3.15.0
my_package==0.0.1